fix(switch): simplify switch implementation + fix reanimated implementation issue
